+++ CHANGELOG.md
@@ -13,6 +13,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 (None)
 
 
+## [0.4.0] - 2020-11-25
+
+### Added
+
+- Added `Calls` methods to `IActor` interface for calling both Tasks and Questions
+
+### Changed
+
+- Replaced `JavaScriptCall` and `JavaScriptElementCall` with `JavaScript`
+
+
 ## [0.3.1] - 2020-11-18
 
 ### Added